+++ reptar/__init__.py
@@ -6,6 +6,7 @@ __version__ = _version.get_versions()["version"]
 
 from .creating import Creator
 from .reptar_file import File
+from .saver import Saver
 from .sampling import Sampler
 
-__all__ = ["Creator", "File", "Sampler"]
+__all__ = ["Creator", "File", "Saver", "Sampler"]